@@ -1,6 +1,6 @@
 module gitlab.com/accumulatenetwork/accumulate
 
-go 1.18
+go 1.19
 
 replace github.com/tendermint/tendermint v0.37.0-rc1 => github.com/AccumulateNetwork/tendermint v0.34.15-0.20221019224404-7f7149107231
 
