@@ -22,7 +22,6 @@ import (
 	"github.com/stretchr/testify/suite"
 	abci "github.com/tendermint/tendermint/abci/types"
 	types2 "github.com/tendermint/tendermint/abci/types"
-	tmed25519 "github.com/tendermint/tendermint/crypto/ed25519"
 	"gitlab.com/accumulatenetwork/accumulate/internal/api/v2/query"
 	"gitlab.com/accumulatenetwork/accumulate/internal/database"
 	"gitlab.com/accumulatenetwork/accumulate/internal/database/indexing"
@@ -1257,7 +1256,7 @@ func TestIssueTokensRefund(t *testing.T) {
 	supplyLimit := big.NewInt(maxSupply)
 	require.NoError(t, acctesting.CreateAdiWithCredits(batch, fooKey, "foo", 1e9))
 	require.NoError(t, acctesting.CreateLiteIdentity(batch, sponsorUrl.String(), 3))
-	require.NoError(t, acctesting.CreateLiteTokenAccount(batch, tmed25519.PrivKey(liteKey), 1e9))
+	require.NoError(t, acctesting.CreateLiteTokenAccount(batch, liteKey, 1e9))
 	require.NoError(t, batch.Commit())
 	liteAddr, err := protocol.LiteTokenAddress(liteKey[32:], "foo.acme/tokens", protocol.SignatureTypeED25519)
 	require.NoError(t, err)
@@ -1355,7 +1354,7 @@ func TestIssueTokensWithSupplyLimit(t *testing.T) {
 	supplyLimit := big.NewInt(maxSupply)
 	require.NoError(t, acctesting.CreateAdiWithCredits(batch, fooKey, "foo", 1e9))
 	require.NoError(t, acctesting.CreateLiteIdentity(batch, sponsorUrl.String(), 3))
-	require.NoError(t, acctesting.CreateLiteTokenAccount(batch, tmed25519.PrivKey(liteKey), 1e9))
+	require.NoError(t, acctesting.CreateLiteTokenAccount(batch, liteKey, 1e9))
 	require.NoError(t, batch.Commit())
 
 	var err error
@@ -1386,8 +1385,8 @@ func TestIssueTokensWithSupplyLimit(t *testing.T) {
 	liteId := liteAcmeAddr.RootIdentity()
 
 	underLimit := int64(1000 * fooPrecision)
-	atLimit := int64(maxSupply - underLimit)
-	overLimit := int64(maxSupply + 1)
+	atLimit := maxSupply - underLimit
+	overLimit := maxSupply + 1
 	// test under the limit
 	n.MustExecuteAndWait(func(send func(*protocol.Envelope)) {
 		body := new(protocol.IssueTokens)
