@@ -220,7 +220,7 @@ func (n *Node) initChain(snapshot ioutil2.SectionReader) ([]byte, error) {
 
 func (n *Node) checkTx(delivery *chain.Delivery, typ abci.CheckTxType) (*protocol.TransactionStatus, error) {
 	// TODO: Maintain a shared batch if typ is not recheck. I tried to do this
-	// but it lead to "attempted to use a commited or discarded batch" panics.
+	// but it lead to "attempted to use a committed or discarded batch" panics.
 
 	batch := n.database.Begin(false)
 	defer batch.Discard()
